+++ py_basic_commands/__init__.py
@@ -6,7 +6,9 @@ from .basic_commands import finput
 from .basic_commands import enter_to_continue
 from .basic_commands import choose_from_list
 from .basic_commands import read_file
+from .basic_commands import write_file
 from .basic_commands import create_file_dir
 from .basic_commands import remove_file_dir
 from .basic_commands import get_dir_path_for_file
+from .basic_commands import fd
 from .basic_commands import join_path